EdkRepo: Sync modifies the global manifest repo if it needs to clone a new repo
In some scenarios, "edkrepo sync -uo" will need to clone a new git repository.
Specifically, if the new version of the project's manifest file adds a git
repository that did not exist before. In that scenario, updating from the
old manifest to the new manifest requires cloning the new git repository.

When sync clones a new git repository, the current implementation will write
a new SourceManifestRepository XML tag to the copy of the project manifest in
the global manifest repository instead of the copy in the workspace. This is
incorrect, the copy in the workspace should be modified, not the copy in the
global manifest repository.

Signed-off-by: Nate DeSimone <nathaniel.l.desimone@intel.com>